move(Gender): Gender 패키지 enums로 이동
