--- src/main/java/soma/ghostrunner/domain/member/Gender.java
@@ -1,5 +1,0 @@
-package soma.ghostrunner.domain.member;
-
-public enum Gender {
-    MALE, FEMALE
-}
