+++ src/main/java/soma/ghostrunner/domain/member/MemberBioInfo.java
@@ -8,7 +8,7 @@ import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.checkerframework.checker.units.qual.A;
+import soma.ghostrunner.domain.member.enums.Gender;
 
 @Embeddable
 @Getter
