+++ src/main/java/soma/ghostrunner/domain/member/enums/Gender.java
@@ -1,0 +1,5 @@
+package soma.ghostrunner.domain.member.enums;
+
+public enum Gender {
+    MALE, FEMALE
+}
